% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkConf}
\alias{checkConf}
\title{Private function for checking R environment for configuration variables.}
\usage{
checkConf()
}
\description{
A variable \code{conf} should be defined in the user's environment as
a character list containing authentication (username, password) as well
as the host address for the desired openscoring server.
}

