% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getModels}
\alias{getModels}
\title{Gets a list of available models on the openscoring server}
\usage{
getModels(host, data.type = "data.frame")
}
\arguments{
\item{host}{the address of the openscoring server from which to get model information}

\item{data.type}{indicates the final R data type for the response (data.frame or list)}
}
\description{
Returns a data.frame with information about the deployed models
}
\keyword{get}
\keyword{list}
\keyword{models}

