% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getModelSchema}
\alias{getModelSchema}
\title{Returns the schema of the given model}
\usage{
getModelSchema(model_name, host)
}
\arguments{
\item{model_name}{the name of the model for which to retrieve schema}
}
\description{
The schema of the model includes "activeFields", "groupFields",
"targetFields", and "outputFields".  Returns a list with this information
associated with the model name.
}
\keyword{get}
\keyword{model}
\keyword{schema}

