% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isDeployed}
\alias{isDeployed}
\title{Tests to see whether a given model is deployed}
\usage{
isDeployed(model_name, host)
}
\arguments{
\item{model_name}{name of the model to test for deployment}

\item{host}{the address of the openscoring server}
}
\description{
Tests to see whether a given model is deployed
}
\keyword{bool}
\keyword{deployed}
\keyword{test}

