% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score}
\alias{predict}
\alias{score}
\title{Scores new data on a deployed model.}
\usage{
score(data, model_name, host)
}
\arguments{
\item{data}{Data on which to execute model.  Must be a data.frame and
the colnames must match the activeFields of model_name.
Run getModelSchema(model_name)$activeFields to get the active}

\item{model_name}{name of the model used to score}

\item{host}{Address of openscoring server (i.e. "http://localhost:8080/openscoring")}
}
\description{
Takes feature data and returns a prediction response by applying
the selected model to the data provided.  This can be called
in either "single prediction" mode (i.e. one observation of features,
one set of prediction values returned), or "batch" mode where
prediction responses are returned for a collection of observations.
}

