% Generated by roxygen2 (4.0.2): do not edit by hand
\name{specialJSON}
\alias{specialJSON}
\title{Private function that returns JSON structure required by openscoring server.}
\usage{
specialJSON(x)
}
\arguments{
\item{x}{data.frame with the data to be parsed into JSON.}
}
\description{
To POST data to be scored, must conform to the JSON structure accepted by the
openscoring server.  Single example, using the first row of the iris dataset:
}

