% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deleteModel}
\alias{deleteModel}
\title{Deletes the model from the openscoring server}
\usage{
deleteModel(model_name, host)
}
\arguments{
\item{model_name}{name of the model to delete}

\item{host}{address of server from which to remove model}
}
\description{
Deletes the model from the openscoring server
}
\keyword{delete}

