% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getModels.r
\name{getModels}
\alias{get}
\alias{getModels}
\alias{models}
\title{Gets a list of available models on the openscoring server.}
\usage{
getModels(host, data.type = "data.frame")
}
\arguments{
\item{host}{the address of the openscoring server from which to get model information}

\item{data.type}{indicates the final R data type for the response (data.frame or list)}
}
\description{
Returns a data.frame with information about the deployed models
}

