% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getModelSchema.r
\name{getModelSchema}
\alias{getModelSchema}
\title{Returns the schema of the given model.}
\usage{
getModelSchema(model_name, host)
}
\arguments{
\item{model_name}{name of the model for which to retrieve schema}
}
\description{
The schema of the model includes "activeFields", "groupFields", "targetFields",
and "outputFields".  \code{getModelSchema} returns a list with this information
associated with the model name.
}

