% Generated by roxygen2 (4.0.2): do not edit by hand
\name{score}
\alias{score}
\title{Score a deployed model on new data}
\usage{
score(model_name, host, data, mode)
}
\arguments{
\item{model_name}{name of the model used to score}

\item{host}{address of openscoring server}

\item{data}{data on which to execute model}

\item{mode}{either 'single' or 'batch'}
}
\description{
Takes feature data and returns a prediction response by applying
the selected model to the data provided.  This can be called
in either "single prediction" mode (i.e. one observation of features,
one set of prediction values returned)
}
\keyword{batch}
\keyword{score}

